backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12

# PDF processing
pymupdf==1.25.2

# Vector search and embeddings
faiss-cpu==1.8.0
sentence-transformers==3.3.1

# Utilities
numpy==1.26.4
